feat: Handle workspace panel hiding event
This commit adds a new `onpopuphidden` attribute to the `<panel>` element that defines the workspace panel. This attribute triggers the `ZenWorkspaces.handlePanelHidden()` function when the panel is hidden.

This change is necessary to allow the application to properly handle the case when the workspace panel is hidden so that the reorder mode can be disabled.

The following changes were made:

- Added the `onpopuphidden` attribute to the `<panel>` element in `zen-popupset.inc.xhtml`.
- Added a new CSS class `.zen-workspace-last-place-drop-target` to `zen-workspaces.css`.
- Modified the `zen-workspaces.css` file to display the `.zen-workspace-last-place-drop-target` class only in reorder mode.